Bug 1818115 - Make FrameId be explicit u64 rather than usize r=gfx-reviewers,bradwerth
Simplifies some upcoming work to change how we store these when
updating primitive dependencies during picture cache updates.

Differential Revision: https://phabricator.services.mozilla.com/D170546

UltraBlame original commit: bdf88c9a1a0690c2598f13e5d93049ad4a886932

+++ gfx/wr/webrender/src/gpu_cache.rs
@@ -47,7 +47,7 @@ pub const GPU_CACHE_INITIAL_HEIGHT: i32 = 20;
 const NEW_ROWS_PER_RESIZE: i32 = 10;
 
 
-const FRAMES_BEFORE_EVICTION: usize = 10;
+const FRAMES_BEFORE_EVICTION: u64 = 10;
